Discard utils.js
